requirements.txt
streamlit>=1.40.0
openai>=1.50.0
python-dotenv>=1.0.0
requests>=2.31.0
duckduckgo-search>=7.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
